Refactor the module to allow for more testing utilities and make it easier to include.
