+++ library/block/block_content_registry/src/testmod/java/org/quiltmc/qsl/block/test/BlockContentRegistryTest.java
@@ -51,8 +51,8 @@ import org.quiltmc.qsl.block.content.registry.api.enchanting.ConstantBooster;
 import org.quiltmc.qsl.block.content.registry.api.enchanting.EnchantingBooster;
 import org.quiltmc.qsl.block.content.registry.api.enchanting.EnchantingBoosterType;
 import org.quiltmc.qsl.block.content.registry.api.enchanting.EnchantingBoosters;
-import org.quiltmc.qsl.game_test.api.QuiltGameTest;
-import org.quiltmc.qsl.game_test.api.QuiltTestContext;
+import org.quiltmc.qsl.testing.api.game.QuiltGameTest;
+import org.quiltmc.qsl.testing.api.game.QuiltTestContext;
 import org.quiltmc.qsl.lifecycle.api.event.ServerWorldTickEvents;
 import org.quiltmc.qsl.registry.attachment.api.RegistryEntryAttachment;
 import org.quiltmc.qsl.registry.attachment.api.RegistryExtensions;
