--- library/core/game_test/build.gradle
@@ -1,25 +1,0 @@
-plugins {
-	id("qsl.module")
-}
-
-qslModule {
-	name = "Quilt Game Test"
-	moduleName = "game_test"
-	id = "quilt_game_test"
-	description = "Utilites to assist in testing the game and mods."
-	library = "core"
-	moduleDependencies {
-		core {
-			api("qsl_base")
-			api("lifecycle_events")
-		}
-		management {
-			testmodOnly("command")
-		}
-	}
-	entrypoints {
-		init {
-			values = ["org.quiltmc.qsl.game_test.impl.QuiltGameTestImpl"]
-		}
-	}
-}
