--- library/core/game_test/src/main/resources/quilt_game_test.mixins.json
@@ -1,21 +1,0 @@
-{
-  "required": true,
-  "package": "org.quiltmc.qsl.game_test.mixin",
-  "compatibilityLevel": "JAVA_17",
-  "mixins": [
-    "ArgumentTypeInfosMixin",
-    "CommandManagerMixin",
-    "MinecraftServerMixin",
-    "StructureTestUtilMixin",
-    "TestCommandMixin",
-    "TestContextAccessor",
-    "TestFunctionsMixin",
-    "TestServerMixin"
-  ],
-  "server": [
-    "server.MainMixin"
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
