+++ library/core/testing/src/main/java/org/quiltmc/qsl/testing/api/package-info.java
@@ -1,0 +1,13 @@
+/**
+ * <h2>The Quilt Testing API</h2>
+ *
+ * The Quilt Testing API is a set of testing utilities to allow modders to test the game and their mods
+ * to ensure that everything is working up to specs.
+ *
+ * The utilities present in this API are:
+ * <ul>
+ *     <li>{@link org.quiltmc.qsl.testing.api.game Game Test API} which allows the usage of the Minecraft's custom automated testing framework.</li>
+ * </ul>
+ */
+
+package org.quiltmc.qsl.testing.api;
