+++ library/core/testing/src/main/resources/assets/quilt_testing/icon.png
