+++ library/core/testing/src/main/resources/data/quilt/game_test/structures/empty.snbt
