+++ library/core/testing/src/main/resources/quilt_testing.mixins.json
@@ -1,0 +1,21 @@
+{
+  "required": true,
+  "package": "org.quiltmc.qsl.testing.mixin",
+  "compatibilityLevel": "JAVA_17",
+  "mixins": [
+    "ArgumentTypeInfosMixin",
+    "CommandManagerMixin",
+    "MinecraftServerMixin",
+    "StructureTestUtilMixin",
+    "TestCommandMixin",
+    "TestContextAccessor",
+    "TestFunctionsMixin",
+    "TestServerMixin"
+  ],
+  "server": [
+    "server.MainMixin"
+  ],
+  "injectors": {
+    "defaultRequire": 1
+  }
+}
